AI/platformerbot-master/platformerbot-master/requirements.txt
numpy
gym
gym-retro
torch
pillow
kivy
kivy_deps.sdl2
kivy_deps.glew
matplotlib
